+++ src/clay_board_style_system/core/node.nim
@@ -37,7 +37,8 @@ type
     arGroup,
     arImage,
     arStaticText,
-    arLink
+    arLink,
+    arSwitch
 
   ElementState* = enum
     esHover,